+++ src/ResourceManager/Websites/Commands.Websites/Cmdlets/DeploymentSlots/NewAzureWebAppSlot.cs
@@ -69,6 +69,9 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.DeploymentSlots
         [ValidateNotNullOrEmpty]
         public string AseResourceGroupName { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
+        public SwitchParameter AsJob { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();