+++ src/ResourceManager/Media/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Current Release
 * Added Location Completer to -Location parameters allowing tab completion through valid Locations
 * Added ResourceGroup Completer to -ResourceGroup parameters allowing tab completion through resource groups in current subscription
+* Obsoleted -Tags in favor of -Tag for Set-AzureRmMediaService and New-AzureRmMediaService
 
 ## Version 0.8.0
 * Add support for online help